added user input for currently of personas
This version is running in first tests on my machine. Now I need to run it on a faster system to see if the changes actually get applied where I want them

Currently: command line prompt for every persona: "What do you want xy to be working on currently"
Input from that should be added under currently in the scratch.json for that persona, so their short-term memory, and hopefully used in the simulation.