added gpt prompt to format the user input
Inspired by plan.py:
if we just add the user input as the string it is to the currently, that gives us a text, that doesn't make much sense.
Maybe we can ask the llm to format it nicely for us/include the task from the user into the currently to sound natural.